Adjusted load conditions and timing, bruteforced panel load on login
Unable to find a nice way to force the panel creation on login, so I just forced the /aca command upon login.

I left in the adjustments to the load order and timing as these had no ill effect and I see no edge case for the slight adjustments I made.